# backend/docs/openapi.yaml
openapi: 3.0.3
info:
  title: BidzPro 2.0 API
  description: API for BidzPro 2.0 AI-Powered Pre-Construction Intelligence Platform
  version: 1.0.0
  contact:
    name: BidzPro API Support
    email: api-support@bidzpro.com
    url: https://developers.bidzpro.com

servers:
  - url: http://localhost:5000/api/v1
    description: Development server
  - url: https://api.bidzpro.com/v1
    description: Production server

security:
  - BearerAuth: []

components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  schemas:
    Error:
      type: object
      properties:
        error:
          type: object
          properties:
            code:
              type: string
              example: INVALID_INPUT
            message:
              type: string
              example: Invalid project dates provided
            details:
              type: object
              additionalProperties: true

    LoginRequest:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          format: password

    LoginResponse:
      type: object
      properties:
        token:
          type: string
        user:
          type: object
          properties:
            id:
              type: integer
            email:
              type: string
            role:
              type: string

    AnalyticsOverview:
      type: object
      properties:
        total_projects:
          type: integer
          example: 150
        active_bids:
          type: integer
          example: 45
        success_rate:
          type: number
          format: float
          example: 0.75
        revenue_forecast:
          type: number
          format: float
          example: 1250000

    TrendData:
      type: object
      properties:
        data:
          type: array
          items:
            type: object
            properties:
              date:
                type: string
                format: date
              value:
                type: number
        metric:
          type: string
          enum: [bids, revenue, success_rate]

    Project:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        status:
          type: string
          enum: [active, completed, pending]
        bid_amount:
          type: number
        created_at:
          type: string
          format: date-time

    ProjectList:
      type: object
      properties:
        projects:
          type: array
          items:
            $ref: '#/components/schemas/Project'
        total:
          type: integer
        page:
          type: integer
        per_page:
          type: integer

    CreateProjectRequest:
      type: object
      required:
        - name
        - bid_amount
      properties:
        name:
          type: string
        description:
          type: string
        bid_amount:
          type: number
        start_date:
          type: string
          format: date
        end_date:
          type: string
          format: date

paths:
  /auth/login:
    post:
      summary: Authenticate user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Successful authentication
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '401':
          description: Authentication failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /analytics/overview:
    get:
      summary: Get analytics overview
      responses:
        '200':
          description: Analytics overview data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AnalyticsOverview'

  /analytics/trends:
    get:
      summary: Get historical trend data
      parameters:
        - name: start_date
          in: query
          schema:
            type: string
            format: date
        - name: end_date
          in: query
          schema:
            type: string
            format: date
        - name: metric
          in: query
          required: true
          schema:
            type: string
            enum: [bids, revenue, success_rate]
      responses:
        '200':
          description: Trend data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TrendData'

  /projects:
    get:
      summary: Get list of projects
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: per_page
          in: query
          schema:
            type: integer
            default: 20
        - name: status
          in: query
          schema:
            type: string
            enum: [active, completed, pending]
      responses:
        '200':
          description: List of projects
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectList'

    post:
      summary: Create a new project
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateProjectRequest'
      responses:
        '201':
          description: Project created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Project'
        '400':
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'